+++ packages/bigcommerce-payments-integration/src/bigcommerce-payments-alternative-methods/bigcommerce-payments-alternative-methods-payment-strategy.spec.ts
@@ -16,6 +16,7 @@ import {
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
 import {
     getBillingAddress,
+    getConfig,
     PaymentIntegrationServiceMock,
 } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
 import { LoadingIndicator } from '@bigcommerce/checkout-sdk/ui';
@@ -635,6 +636,103 @@ describe('BigCommercePaymentsAlternativeMethodsPaymentStrategy', () => {
     describe('#finalize()', () => {
         it('throws error to inform that order finalization is not required', async () => {
             await expect(strategy.finalize()).rejects.toThrow(OrderFinalizationNotRequiredError);
+        });
+    });
+
+    describe('#polling mechanism', () => {
+        beforeEach(() => {
+            const storeConfigMock = {
+                ...getConfig().storeConfig,
+                checkoutSettings: {
+                    ...getConfig().storeConfig.checkoutSettings,
+                    features: {
+                        'PAYPAL-5624.bcp_ideal_polling': true,
+                    },
+                },
+            };
+
+            jest.spyOn(paymentIntegrationService, 'submitPayment').mockReturnValueOnce(
+                Promise.resolve(paymentIntegrationService.getState()),
+            );
+            jest.spyOn(
+                paymentIntegrationService.getState(),
+                'getStoreConfigOrThrow',
+            ).mockReturnValue(storeConfigMock);
+        });
+
+        it('initialize polling mechanism', async () => {
+            jest.spyOn(bigCommercePaymentsIntegrationService, 'getOrderStatus').mockResolvedValue(
+                PayPalOrderStatus.Approved,
+            );
+
+            const payload = {
+                payment: {
+                    methodId: 'ideal',
+                    gatewayId: 'bigcommerce_payments_apms',
+                },
+            };
+
+            await strategy.initialize(initializationOptions);
+
+            eventEmitter.emit('onApprove');
+
+            await strategy.execute(payload);
+
+            await new Promise((resolve) => process.nextTick(resolve));
+
+            expect(bigCommercePaymentsIntegrationService.getOrderStatus).toHaveBeenCalled();
+        });
+
+        it('request order status with proper payload', async () => {
+            jest.spyOn(bigCommercePaymentsIntegrationService, 'getOrderStatus').mockResolvedValue(
+                PayPalOrderStatus.Approved,
+            );
+
+            const payload = {
+                payment: {
+                    methodId: 'ideal',
+                    gatewayId: 'bigcommerce_payments_apms',
+                },
+            };
+
+            await strategy.initialize(initializationOptions);
+
+            eventEmitter.emit('onApprove');
+
+            await strategy.execute(payload);
+
+            await new Promise((resolve) => process.nextTick(resolve));
+
+            expect(bigCommercePaymentsIntegrationService.getOrderStatus).toHaveBeenCalledWith(
+                'bigcommerce_payments_apms',
+            );
+        });
+
+        it('deinitialize polling mechanism', async () => {
+            jest.spyOn(bigCommercePaymentsIntegrationService, 'getOrderStatus').mockResolvedValue(
+                PayPalOrderStatus.Approved,
+            );
+
+            const payload = {
+                payment: {
+                    methodId: 'ideal',
+                    gatewayId: 'bigcommerce_payments_apms',
+                },
+            };
+
+            jest.spyOn(global, 'clearTimeout');
+
+            await strategy.initialize(initializationOptions);
+
+            eventEmitter.emit('onApprove');
+
+            await strategy.execute(payload);
+
+            await new Promise((resolve) => process.nextTick(resolve));
+
+            await strategy.deinitialize();
+
+            expect(clearTimeout).toHaveBeenCalled();
         });
     });
 });